pyproject.toml
[tool.poetry]
name = "mxcubecore"
version = "1.0.0"
license = "LGPL-3.0-or-later"
description = "Core libraries for the MXCuBE application"
authors = ["The MXCuBE collaboration <mxcube@esrf.fr>"]
maintainers = [
    "Marcus Oscarsson <mxcube@esrf.fr>",
    "Rasmus H. Fogh <rhfogh@globalphasing.com>",
    "Antonia Beteva <beteva@esrf.fr>",
]
readme = "README.md"
homepage = "http://github.com/mxcube/mxcubecore"
repository = "http://github.com/mxcube/mxcubecore"
documentation = ""
keywords = ["mxcube", "mxcube3", "mxcubecore"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
gevent = "^21.12.0"
greenlet = "^1.1.3"
jsonpickle = "^2.2.0"
jsonschema = "^4.17.1"
lxml = "^4.9.1"
matplotlib = [ { version = "3.5.3", python = "3.7" }, { version = "^3.6.2", python = ">=3.8" } ]
numpy = [ { version = "1.21.6", python = "3.7" }, { version = "^1.23.5", python = ">=3.8" } ]
Pillow = "^9.3.0"
pydantic = "^1.10.2"
PyDispatcher = "^2.0.6"
"ruamel.yaml" = "^0.17.21"
scipy = [ { version = "1.7.3", python = "3.7" }, { version = "^1.9.3", python = ">=3.8" } ]
suds-py3 = "1.4.5.0"
psutil = "^5.9.4"
gipc = "^1.4.0"
importlib_metadata = { version = ">=4.12.0", python = "3.7" }
py4j = "0.10.9.7"
f90nml = "1.4.3"
lucid3 = { git = "https://github.com/mxcube/lucid3.git", branch = "master"}
PyTango = { version = "^9.3.6", optional = true }
python-ldap = "^3.4.3"
requests = "^2.28.1"
colorama = "^0.4.6"

[tool.poetry.dev-dependencies]
pylint = [ { version = "==2.13.9", python = "<=3.7.1" }, { version = "^2.15.6", python = ">=3.8" } ]
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[tool.poetry.extras]
tango = ["PyTango"]


[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "poetry.core.masonry.api"
